+++ R/deprecated.R
@@ -23,7 +23,7 @@
 cff_extract_to_bibtex <- function(x,
                                   what = c("preferred", "references", "all")) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "0.5.0", "cff_extract_to_bibtex()",
       details = "Function renamed, use `as_bibentry()` instead."
     )
@@ -37,7 +37,7 @@ cff_extract_to_bibtex <- function(x,
 cff_to_bibtex <- function(x,
                           what = c("preferred", "references", "all")) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "cff_extract_to_bibtex()",
       details = "Function renamed, use `as_bibentry()` instead."
     )
@@ -103,7 +103,7 @@ cff_to_bibtex <- function(x,
 cff_from_bibtex <- function(x, encoding = "UTF-8", ...) {
   if (length(x) == 1 && file_exist_abort(x)) {
     if (requireNamespace("lifecycle", quietly = TRUE)) {
-      lifecycle::deprecate_soft(
+      lifecycle::deprecate_warn(
         "1.0.0", "cff_from_bibtex()", "cff_read_bib()"
       )
     }
@@ -113,7 +113,7 @@ cff_from_bibtex <- function(x, encoding = "UTF-8", ...) {
   }
 
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "cff_from_bibtex()", "cff_read_bib_text()"
     )
   }
@@ -162,7 +162,7 @@ write_bib <- function(x,
                       verbose = TRUE,
                       ascii = FALSE) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "write_bib()", "cff_write_bib()"
     )
   }
@@ -178,7 +178,7 @@ write_citation <- function(x,
                            verbose = TRUE,
                            ...) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "write_citation()", "cff_write_citation()"
     )
   }
@@ -245,7 +245,7 @@ write_citation <- function(x,
 #' as_cff_person("Herbert von Karajan")
 cff_parse_person <- function(person) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "cff_parse_person()", "as_cff_person()"
     )
   }
@@ -257,7 +257,7 @@ cff_parse_person <- function(person) {
 #'
 cff_parse_person_bibtex <- function(person) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "cff_parse_person_bibtex()", "as_cff_person()"
     )
   }
@@ -292,7 +292,7 @@ cff_parse_person_bibtex <- function(person) {
 #'
 cff_parse_citation <- function(bib) {
   if (requireNamespace("lifecycle", quietly = TRUE)) {
-    lifecycle::deprecate_soft(
+    lifecycle::deprecate_warn(
       "1.0.0", "cff_parse_citation()", "as_cff.bibentry()"
     )
   }