+++ NEWS.md
@@ -1,4 +1,4 @@
-# cffr (development version)
+# cffr 1.1.0
 
 -   Now **cffr** adds automatically dois to packages on **CRAN** (e.g.
     <https://doi.org/10.32614/CRAN.package.cffr>):
@@ -10,6 +10,7 @@
         the package dependency is on **CRAN**, the **CRAN** doi would be used
         for the dependency in the `references` key.
 -   Extract `commit` from `RemoteSha`.
+-   Update `cran_to_spdx` dataset.
 
 # cffr 1.0.1